moved SyncedFlushResponse to org.elasticsearch.client
-- Also Fixed indentation
