+++ client/rest-high-level/src/main/java/org/elasticsearch/client/IndicesClient.java
@@ -54,7 +54,6 @@ import org.elasticsearch.action.admin.indices.shrink.ResizeRequest;
 import org.elasticsearch.action.admin.indices.shrink.ResizeResponse;
 import org.elasticsearch.action.admin.indices.template.put.PutIndexTemplateRequest;
 import org.elasticsearch.action.admin.indices.template.put.PutIndexTemplateResponse;
-import org.elasticsearch.client.response.indices.flush.SyncedFlushResponse;
 
 import java.io.IOException;
 import java.util.Collections;
