+++ client/rest-high-level/src/test/java/org/elasticsearch/client/SyncedFlushResponseTests.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.elasticsearch.client.response.indices.flush;
+package org.elasticsearch.client;
 
 import java.io.IOException;
 import java.util.Map;
